# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Human Service API",
    "description": "Scrapes and caches a person's online presence (websites, LinkedIn, blog, etc.) to provide writing context for AI content generation.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
      "UpsertHumanResponse": {
        "type": "object",
        "properties": {
          "human": {
            "$ref": "#/components/schemas/Human"
          },
          "created": {
            "type": "boolean"
          }
        },
        "required": [
          "human",
          "created"
        ]
      },
      "Human": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "expertise": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "knownFor": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "bio",
          "expertise",
          "knownFor",
          "imageUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "UpsertHumanRequest": {
        "type": "object",
        "properties": {
          "appId": {
            "type": "string",
            "minLength": 1
          },
          "orgId": {
            "type": "string",
            "minLength": 1
          },
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[a-z0-9-]+$"
          },
          "urls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "minItems": 1
          },
          "bio": {
            "type": "string"
          },
          "expertise": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "knownFor": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "maxPages": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          }
        },
        "required": [
          "appId",
          "orgId",
          "userId",
          "name",
          "slug",
          "urls"
        ]
      },
      "ListHumansResponse": {
        "type": "object",
        "properties": {
          "humans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Human"
            }
          }
        },
        "required": [
          "humans"
        ]
      },
      "GetHumanResponse": {
        "type": "object",
        "properties": {
          "human": {
            "$ref": "#/components/schemas/Human"
          }
        },
        "required": [
          "human"
        ]
      },
      "GetMethodologyResponse": {
        "type": "object",
        "properties": {
          "methodology": {
            "$ref": "#/components/schemas/Methodology"
          },
          "isExpired": {
            "type": "boolean"
          }
        },
        "required": [
          "methodology"
        ]
      },
      "Methodology": {
        "type": "object",
        "properties": {
          "humanId": {
            "type": "string",
            "format": "uuid"
          },
          "frameworks": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Framework"
            }
          },
          "strategicPatterns": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "toneOfVoice": {
            "$ref": "#/components/schemas/ToneProfile"
          },
          "persuasionStyle": {
            "$ref": "#/components/schemas/PersuasionStyle"
          },
          "contentSignatures": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "avoids": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "extractionModel": {
            "type": "string",
            "nullable": true
          },
          "extractedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "humanId",
          "frameworks",
          "strategicPatterns",
          "toneOfVoice",
          "persuasionStyle",
          "contentSignatures",
          "avoids",
          "extractionModel",
          "extractedAt"
        ]
      },
      "Framework": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "applicationContext": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "applicationContext"
        ]
      },
      "ToneProfile": {
        "type": "object",
        "nullable": true,
        "properties": {
          "register": {
            "type": "string"
          },
          "pace": {
            "type": "string"
          },
          "vocabulary": {
            "type": "string"
          },
          "perspective": {
            "type": "string"
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "register",
          "pace",
          "vocabulary",
          "perspective",
          "examples"
        ]
      },
      "PersuasionStyle": {
        "type": "object",
        "nullable": true,
        "properties": {
          "primary": {
            "type": "string"
          },
          "techniques": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "callToAction": {
            "type": "string"
          }
        },
        "required": [
          "primary",
          "techniques",
          "callToAction"
        ]
      },
      "ExtractResponse": {
        "type": "object",
        "properties": {
          "human": {
            "$ref": "#/components/schemas/Human"
          },
          "methodology": {
            "$ref": "#/components/schemas/Methodology"
          },
          "pagesScraped": {
            "type": "integer"
          }
        },
        "required": [
          "human",
          "methodology",
          "pagesScraped"
        ]
      },
      "ExtractRequest": {
        "type": "object",
        "properties": {
          "appId": {
            "type": "string",
            "minLength": 1
          },
          "orgId": {
            "type": "string",
            "minLength": 1
          },
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "keySource": {
            "type": "string",
            "enum": [
              "app",
              "byok",
              "platform"
            ]
          },
          "runId": {
            "type": "string"
          },
          "forceRefresh": {
            "type": "boolean"
          }
        },
        "required": [
          "appId",
          "orgId",
          "userId",
          "keySource"
        ]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "service": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "service"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/humans": {
      "post": {
        "summary": "Create or update a human expert",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertHumanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Human created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertHumanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List humans for an org",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "appId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "orgId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Humans found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListHumansResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/humans/{id}": {
      "get": {
        "summary": "Get human by ID",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Human found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHumanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Human not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/humans/{id}/methodology": {
      "get": {
        "summary": "Get cached methodology for a human",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Methodology found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMethodologyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Methodology not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/humans/{id}/extract": {
      "post": {
        "summary": "Trigger scrape and AI methodology extraction",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extraction completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Human not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}